Swapped from Room to SQLDelight for database access

+++ app/src/main/java/com/sidharth/mosam/data/remote/WeatherResponse.kt
@@ -19,17 +19,17 @@ data class CurrentWeather(
     val temp: Double,
     @SerializedName("feels_like")
     val feelsLike: Double,
-    val pressure: Int,
-    val humidity: Int,
+    val pressure: Long,
+    val humidity: Long,
     @SerializedName("dew_point")
     val dewPoint: Double,
     val uvi: Double,
-    val clouds: Int,
-    val visibility: Int,
+    val clouds: Long,
+    val visibility: Long,
     @SerializedName("wind_speed")
     val windSpeed: Double,
     @SerializedName("wind_deg")
-    val windDeg: Int,
+    val windDeg: Long,
     @SerializedName("wind_gust")
     val windGust: Double,
     val weather: List<Weather>
